Fixed size related variables.

+++ FitAlgo/include/hole.h
@@ -1,10 +1,11 @@
 #ifndef _MA_HOLE_H
 #define _MA_HOLE_H
 
-struct Hole
-{	
-	int size;
-	int startAddress;
+#include<stddef.h>
+
+struct Hole{
+	size_t size;
+	size_t startAddress;
 	struct Hole *next;
 	struct Hole *previous;
 };
