+++ FitAlgo/include/object.h
@@ -1,10 +1,12 @@
 #ifndef _MA_OBJECT_H
 #define _MA_OBJECT_H
 
+#include<stddef.h>
+
 struct Object
 {
-    int size;
-    int startAddress;
+    size_t size;
+    size_t startAddress;
     struct Object *next;
     struct Object *previous;
 };
